Renamed BS and updated PR (get rid of n19)
