+++ packages/nexus-bridge/SIEBEL/SIF/Nexus BS.sif
@@ -26,7 +26,7 @@
 		   EXTERNAL_USE="N"
 		   HIDDEN="N"
 		   INACTIVE="N"
-		   NAME="N19 BS"
+		   NAME="Nexus BS"
 		   SERVER_ENABLED="N"
 		   STATE_MANAGEMENT_TYPE="Stateful"
 		   WEB_SERVICE_ENABLED="N"
