+++ packages/nexus-bridge/src/NexusLocaleData.js
